Revert "openjdk: Fix QEMU call to use sysroot for libs"
This reverts commit ab51e51d9a9a51bedd19de09bcba4ce830dfc5e7.

Using an oe setup with ptest enabled makes qemu never to return due to testfiles
staged by udev in the sysroot
See also:
http://lists.openembedded.org/pipermail/openembedded-core/2014-September/097098.html

All binaries compiled for qemu are now linked with -static. The -L, -E parameters
pointing to the sysroot and dynamic linker are not needed anymore and removel works
around the issue.

Further details:
http://lists.openembedded.org/pipermail/openembedded-devel/2015-January/099795.html

Signed-off-by: Max Krummenacher <max.oss.09@gmail.com>
Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>