cacao-initial-native: Fix sed matching
The recipe uses sed to patch src/scripts/java.in so that it execs cacao-initial
instead of cacao. However, if configure step is run again, -initial is appended
again to the string, leading to 'exec cacao-initial-initial'. This causes the
build to fail.

This is fixed by changing the sed command to consider larger portion of the
source string. This ensures that the -initial is only appended once.

Signed-off-by: Erkka Kääriä <erkka.kaaria@intel.com>
Signed-off-by: Maxin B. John <maxin.john@intel.com>