ecj-initial.in: Disable multithreading to mitigate segfaulting
Cacao-initial-native suffers from sporadic segmentation faults, when the Java
program uses multiple Java threads, due to a race condition. In particular,
jamvm-native and classpath-native builds would occasionally fail as the
eclipse compiler used during build task by default uses multiple threads.

As fixing the race condition would require large changes to the
cacao-initial-native codebase, and as it is only used briefly during the
bootstrapping process, it is easier and safer to just force ecj-initial to run
in a singlethreaded mode.

Signed-off-by: Erkka Kääriä <erkka.kaaria@intel.com>
Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>