openjdk-7: drop an inappropriate patch
The patch icedtea-jdk-build-nio-gensor-genuc-gensc-static.patch
is to add static option to build nio gensor/genuc/gensc, and
was added when upgrading to IcedTea 2.6.1 OpenJDK 7u85b01.

Another patch icedtea-jdk-nio-use-host-cc.patch uses host cc
for nio build so it fails on host that lack of static libraries:
| /usr/bin/ld: cannot find -lpthread

And there is no comments explaining why it needs static linking,
so drop the patch for now unless we have better understanding
and make a more proper patch.

Signed-off-by: Jackie Huang <jackie.huang@windriver.com>
Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>