Openjdk-8: Fix build flags for adlc
Adlc is a native tool that openjdk builds and uses during its build process.
Bitbake however passes target machine specific CFLAGS and CXXFLAGS because we
are crosscompiling openjdk for the target architecture. This can cause issues,
if these flags contains values, that work for the Yocto provided crosscompiler
but not for the system compiler.

As an example, compilation will fail on Ubuntu 14.04 if
-fstack-protector-strong is specified in the distro security flags. Ubuntu
14.04 ships with GCC 4.8, whereas this flag is only supported by GCC 4.9+.

Signed-off-by: Erkka Kääriä <erkka.kaaria@intel.com>
Signed-off-by: Maxin B. John <maxin.john@intel.com>