icedtea7-native: Fix IOException constructor compile error
Under some circumstances, the bootstrap error fails with the following
error:

build-bootstrap-javac:
 [javac] Compiling 1 source file to
 build/openjdk.build-boot/langtools/build/bootstrap/classes
 [javac] 1. ERROR in /com/sun/tools/javac/file/ZipFileIndex.java
 [javac]     super(message, cause);
 [javac] The constructor IOException(String, Throwable) is undefined

The reason for that error is that rt.jar does not contain the openjdk
classes but the ones from classpath 0.99.

The error eventually goes away after a couple of hours :S.

After some tedious debug, I figured out that the problem was the
timestamp handling of zip/unzip.

The Makefile launches zip with the following options:

$(ZIP) -qur

which only updates the content of the classes that have a timestamp
older than the new ones.

According to zip manpage:
"""
Note that under many operating systems, the TZ (timezone) environment
variable must be set correctly in order for -f and -u to work properly
"""

On this patch, I remove the -u option from zip, always replacing the
classes inside rt.jar with the classes from openjdk. I cannot think of
an scenario where this may fail.

Signed-off-by: Ricardo Ribalda Delgado <ricardo.ribalda@gmail.com>
Signed-off-by: Maxin B. John <maxin.john@intel.com>