openjdk-8: rework do_patch (pt 2 - use bitbake variable overrides)
This currently uses a hand-crafted solution to distinguish
between common, native-only and and target-only patches.
That is a bit hard to follow as patches are being applied in
strange order, and is also non-standard.

Instead, we can just use bitbake variable overrides. This
makes it much easier to work with this recipe, as:
* it is clear in which order patches are going to be applied by
  looking at the recipe
* it is clear which patches are meant to be common, for build,
  or target
* old patches that are still lying around in WORKDIR (e.g.
  because rm_work is not enabled), but that have been removed
  from SRC_URI are no longer incorrectly applied
* if patches fail to apply, we know exactly which patch has
  failed
* we can use PATCHTOOL = without any ill effects

Signed-off-by: André Draszik <andre.draszik@jci.com>
Signed-off-by: Maxin B. John <maxin.john@intel.com>