openjdk-7: clarify a bitbake warning
We get a bitbake warning during recipe building complaining about
unsupported architectures unconditionally. That check is relevant
only for shark builds, so it is quite confusing for non-shark
builds.

Make the warning conditional on whether shark builds are enabled
or not.

This is the same patch as the one for openjdk-8 by André Draszik
(commit 86c729cb51f880fd5a1ec6485baddfa2bedaa998)

Signed-off-by: Richard Leitner <richard.leitner@skidata.com>
Acked-by: Henning Heinold <henning@itconsulting-heinold.de>