openjdk: remove incorrect PROVIDES (java2-vm java2-runtime)
There are two issues here:
* PROVIDES is not package specific
* it doesn't look like the intention is to be able to
  DEPENDS on java2-vm or java2-runtime anyway

Regarding java2-vm:
-------------------
java2-vm was originally used in the OpenJDK 6 & 7 recipes to be able to
select shark, zero, cacao, or jamvm VMs. OpenJDK-6 is not available any
more, and OpenJDK-7 has removed support for most of this in commit
38f4c1365c11 ("openjdk7: remove broken/unsupported VM's")

It is not clear why it was added to the OpenJDK-8 recipe either.

Given OpenJDK-7 has no way of using the VM compiled as part of OpenJDK-8,
and given that no part of the OpenJDK-8 makes use of the java2-vm part, the
correct solution here is to actually *remove* the incorrect PROVIDES as
well as *R*PROVIDES statements for java2-vm completely.

Regarding java2-runtime:
------------------------
Again, looking at the other uses of this:
java2-runtime is a virtual runtime package name, which is provided by
different *runtime* packages (created by the OpenJDK-7, OpenJDK-8, JamVM, or
Cacao recipes). Other recipes only ever RDEPEND on java2-runtime. It makes
no sense for the OpenJDK-8 recipe to PROVIDES java2-runtime given the above.

Remove the incorrect (R)PROVIDES statements.

Signed-off-by: André Draszik <andre.draszik@jci.com>
Signed-off-by: Richard Leitner <richard.leitner@skidata.com>