java-library: make packages overriding PACKAGE_ARCH work again
Recent changes is OE have caused the traditional approach of inheriting
allarch and setting PACKAGE_ARCH not work anymore. Once allarch is
inherited, PACKAGE_ARCH can not be overridden afterwards.
See commit a23c482cab4f ("allarch: only enable allarch when multilib
is not used") d9ba0219b2f6 in poky.
    http://git.openembedded.org/openembedded-core/commit/?id=a23c482cab4f874f4a6a6889716123569eb5ece9

The error manifests itself with configure trying to
   --host=allarch-poky-linux --target=allarch-poky-linux
which fails.

To work around this we can make java-library's allarch inherit
conditional, as is done e.g. in OE-core for packagegroup.bbclass
    http://git.openembedded.org/openembedded-core/commit/?id=9c826962ec8fa45c2b035427442b90a41517144e
    http://git.openembedded.org/openembedded-core/commit/?id=2c9b1d304daade7b0907320aeb9c522e7ab9dcab

So this commit does exactly that, and fixes the two users of this to
follow the new approach.

Signed-off-by: André Draszik <andre.draszik@jci.com>
Signed-off-by: Richard Leitner <richard.leitner@skidata.com>