icedtea7-native: set default x11 PACKAGECONFIG from DISTRO_FEATURES
If x11 is specified in DISTRO_FEATURES, then make it a default
PACKAGECONFIG option when building icedtea7-native; otherwise, the
libx11-native dependency won't be satisfied.

The recipe reads as though icedtea7-native can build completely without
X11, but if the libx11-native headers are not installed on the host
system, compile errors in multiple units because of a missing X11/Xlib.h.

In file included from ../../../src/share/native/sun/awt/../java2d/pipe/Region.h:34:0,
                 from ../../../src/share/native/sun/awt/../java2d/loops/GraphicsPrimitiveMgr.c:31:
../../../src/solaris/native/sun/awt/utility/rect.h:31:22: fatal error: X11/Xlib.h: No such file or directory

In this example, it appears that GraphicsPrimitiveMgr.c is supposed to
be compiled even in headless mode, so the dependency still seems to be
needed.

Signed-off-by: Kyle Russell <bkylerussell@gmail.com>
Signed-off-by: Richard Leitner <richard.leitner@skidata.com>