openjdk-8: Workaround TOPDIR not getting expanded in configure.ac
Somehow the TOPDIR environment doesn't get expanded in configure.ac. Suspecting
a clash with OE's internal TOPDIR variable, I tried replacing it with JDKTOPDIR
but that resulted in the same error.

| autoreconf: configure.ac: creating directory $TOPDIR/common/autoconf/build-aux
| autoreconf: error: cannot create $TOPDIR/common/autoconf/build-aux: No such file or directory

The workaround implemented here is to replace $TOPDIR in the file by its assigned
value ${S}. This makes the error go away and the native build succeed.

Signed-off-by: Mike Looijmans <mike.looijmans@topic.nl>
Signed-off-by: Richard Leitner <richard.leitner@skidata.com>