openjdk-build-helper: Use python raw string for regex pattern
See https://docs.python.org/3/library/re.html

Invalid escape sequences in Python’s usage of the backslash in string literals
now generate a DeprecationWarning, and in the future this will become a
SyntaxError. The solution is to use Python’s raw string notation for regular
expression patterns; backslashes are not handled in any special way in a string
literal prefixed with 'r'.

Signed-off-by: Douglas Royds <douglas.royds@taitradio.com>
Signed-off-by: Richard Leitner <richard.leitner@skidata.com>