README: remove Maxin John from Maintainers
As messages to the given e-mail address of Maxin John
(maxin.john@intel.com) are undeliverable remove the entry from the
README file.

Maxin, if you read this and still want to help maintaining meta-java
please drop me a line! It would be great to have you back on board.

Signed-off-by: Richard Leitner <richard.leitner@skidata.com>