dbus-java: Include features check from classpath
classpath is an indirect dependency of this package that also has a
required distro feature on x11. Without having the same feature check
here, a world build would fail when x11 is not enabled. Reflecting the
same feature check in this recipe, makes sure that world will skip it as
well.

Signed-off-by: Andrei Gherzan <andrei.gherzan@huawei.com>
Signed-off-by: Richard Leitner <richard.leitner@skidata.com>