conf: layer: add free(1) to HOSTTOOLS_NONFATAL for icedtea7-native
icedtea7-native build uses free -m (from the procfs suite of tools) to
determine memory bound during compile. This always fails for us,
because free is not passed through from the host environment.
This results in a few warnings in the logs:

  /bin/sh: line 1: free: command not found
    [..snip..]/Sanity-Settings.gmk:195: WARNING: MB_OF_MEMORY should not
    be empty [Sanity-Settings.gmk]

But otherwise this is handled gracefully. Problems arise when building
icedtea7-native on hosts with more modest RAM sizes: Build process
will consume so much RAM that the OOM killer is invoked. Let us avoid
this by passing through free from the host environment.

We use HOSTTOOLS_NONFATAL instead of HOSTTOOLS as to not break
currently working setups that have enough RAM but lack free(1).
This needs to be done globally due to the way HOSTTOOLS_NONFATAL is
evaluated, we thus stick it into the layer.conf.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Signed-off-by: Richard Leitner <richard.leitner@skidata.com>